+++ fighters/common/src/opff/other.rs
@@ -332,6 +332,38 @@ pub extern "C" fn left_stick_flick_counter(fighter: &mut L2CFighterCommon) {
     }
 }
 
+const HANDLE: i32 = 0x01FF;
+const COUNTER: i32 = 0x01FE;
+
+unsafe extern "C" fn kill_screen_handler(fighter: &mut L2CFighterCommon) {
+    if VarModule::get_int(fighter.object(), COUNTER) > 0 {
+        let scale = (30 - VarModule::get_int(fighter.object(), COUNTER)) as f32 / 30.0 * 5.0;
+        EffectModule::set_scale(fighter.module_accessor, VarModule::get_int(fighter.object(), HANDLE) as u32, &Vector3f::new(scale, 1.0, scale));
+        if VarModule::get_int(fighter.object(), COUNTER) == 10 {
+            fighter.clear_lua_stack();
+            lua_args!(fighter, Hash40::new("sys_bg_finishhit"), false, false);
+            smash::app::sv_animcmd::EFFECT_OFF_KIND(fighter.lua_state_agent);
+            fighter.pop_lua_stack(1);
+            fighter.clear_lua_stack();
+            lua_args!(fighter, Hash40::new("sys_bg_black"), false, false);
+            smash::app::sv_animcmd::EFFECT_OFF_KIND(fighter.lua_state_agent);
+            fighter.pop_lua_stack(1);
+        }
+        if VarModule::get_int(fighter.object(), COUNTER) == 5 {
+            SlowModule::clear_whole(fighter.boma());
+        }
+        VarModule::dec_int(fighter.object(), COUNTER);
+    } else {
+        VarModule::set_int(fighter.object(), HANDLE, 0);
+    }
+    
+    // let frame = WorkModule::get_float(fighter.module_accessor, *FIGHTER_INSTANCE_WORK_ID_FLOAT_DAMAGE_REACTION_FRAME);
+    // let last = WorkModule::get_float(fighter.module_accessor, *FIGHTER_INSTANCE_WORK_ID_FLOAT_DAMAGE_REACTION_FRAME_LAST);
+    // if frame - 1.0 < 0.0  && frame != 0.0 {
+    //     println!("<{:.2}, {:.2}>", PostureModule::pos_x(fighter.module_accessor), PostureModule::pos_y(fighter.module_accessor));
+    // }
+}
+
 pub unsafe fn run(fighter: &mut L2CFighterCommon, boma: &mut BattleObjectModuleAccessor, cat: [i32 ; 4], status_kind: i32, situation_kind: i32, fighter_kind: i32, stick_x: f32, stick_y: f32, facing: f32) {
     airdodge_refresh_on_hit_disable(boma, status_kind);
     suicide_throw_mashout(fighter, boma);
@@ -340,5 +372,6 @@ pub unsafe fn run(fighter: &mut L2CFighterCommon, boma: &mut BattleObjectModuleA
     faf_ac_debug(fighter);
     taunt_parry_forgiveness(fighter);
     custom_dash_anim_support(fighter);
+    kill_screen_handler(fighter);
 }
