+++ fighters/pikachu/Cargo.toml
@@ -11,4 +11,3 @@ skyline = "0.2.1"
 skyline_smash = { git = "https://github.com/blu-dev/skyline-smash", features = ["weak_l2cvalue"] }
 smash_script = { git = "https://github.com/blu-dev/smash-script", branch = "development" }
 smashline = { git = "https://github.com/HDR-Development/smashline" }
-smash_rs = { package = "smash", git = "https://github.com/blu-dev/smash-rs" }
